# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: DEVLINK API
  description: API service for the Distributed Link Intelligence Platform.
  version: 0.1.0
paths:
...
  /api/v1/links:
    post:
      tags:
      - links
      summary: Create a short link
      description: Create a short link for the authenticated user. Duplicate normalized
        URLs are deduplicated unless a custom alias is requested.
      operationId: create_link_api_v1_links_post
      security:
      - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkResponse'
        '400':
          description: Invalid alias, URL, expiration, or request payload.
        '401':
          description: Authentication is required.
        '403':
          description: The current user cannot access this link/workspace.
        '404':
          description: Link not found.
        '409':
          description: Custom alias already exists.
        '410':
          description: The public link is inactive or expired.
        '422':
          description: Request validation failed.
        '429':
          description: Rate limit exceeded.
    get:
      tags:
      - links
      summary: List links
      description: List links visible to the authenticated user with offset pagination.
      operationId: list_links_api_v1_links_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: include_inactive
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Include Inactive
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LinkResponse'
                title: Response List Links Api V1 Links Get
        '400':
          description: Invalid alias, URL, expiration, or request payload.
        '401':
          description: Authentication is required.
        '403':
          description: The current user cannot access this link/workspace.
        '404':
          description: Link not found.
        '409':
          description: Custom alias already exists.
        '410':
          description: The public link is inactive or expired.
        '422':
          description: Request validation failed.
        '429':
          description: Rate limit exceeded.
  /api/v1/links/bulk:
    post:
      tags:
      - links
      summary: Bulk-create short links
      description: Shorten up to 50 URLs and return per-item success/error results.
        Partial failures do not fail the whole request.
      operationId: bulk_create_links_api_v1_links_bulk_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkLinkCreateRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkLinkCreateResponse'
        '400':
          description: Invalid alias, URL, expiration, or request payload.
        '401':
          description: Authentication is required.
        '403':
          description: The current user cannot access this link/workspace.
        '404':
          description: Link not found.
        '409':
          description: Custom alias already exists.
        '410':
          description: The public link is inactive or expired.
        '422':
          description: Request validation failed.
        '429':
          description: Rate limit exceeded.
      security:
      - OAuth2PasswordBearer: []
  /api/v1/links/{short_code}/qr:
    get:
      tags:
      - links
      summary: Get link QR code
      description: Return a cached PNG or SVG QR code for a public redirect URL. This
        endpoint is public but only returns QR codes for available links.
      operationId: get_link_qr_api_v1_links__short_code__qr_get
      parameters:
      - name: short_code
        in: path
        required: true
        schema:
          type: string
          title: Short Code
      - name: image_format
        in: query
        required: false
        schema:
          type: string
          pattern: ^(png|svg)$
          default: png
          title: Image Format
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '400':
          description: Invalid alias, URL, expiration, or request payload.
        '401':
          description: Authentication is required.
        '403':
          description: The current user cannot access this link/workspace.
        '404':
          description: Link not found.
        '409':
          description: Custom alias already exists.
        '410':
          description: The public link is inactive or expired.
        '422':
          description: Request validation failed.
        '429':
          description: Rate limit exceeded.
  /api/v1/links/{link_id}:
    get:
      tags:
      - links
      summary: Get a link
      description: Return one link by numeric ID if the authenticated user can access
        it.
      operationId: get_link_api_v1_links__link_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: link_id
        in: path
        required: true
        schema:
          type: integer
          title: Link Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkResponse'
        '400':
          description: Invalid alias, URL, expiration, or request payload.
        '401':
          description: Authentication is required.
        '403':
          description: The current user cannot access this link/workspace.
        '404':
          description: Link not found.
        '409':
          description: Custom alias already exists.
        '410':
          description: The public link is inactive or expired.
        '422':
          description: Request validation failed.
        '429':
          description: Rate limit exceeded.
    patch:
      tags:
      - links
      summary: Update a link
      description: Update link metadata, destination, alias, password, active state,
        or expiration. Destination changes trigger a new asynchronous safety check.
      operationId: update_link_api_v1_links__link_id__patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: link_id
        in: path
        required: true
        schema:
          type: integer
          title: Link Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkResponse'
        '400':
          description: Invalid alias, URL, expiration, or request payload.
        '401':
          description: Authentication is required.
        '403':
          description: The current user cannot access this link/workspace.
        '404':
          description: Link not found.
        '409':
          description: Custom alias already exists.
        '410':
          description: The public link is inactive or expired.
        '422':
          description: Request validation failed.
        '429':
          description: Rate limit exceeded.
    delete:
      tags:
      - links
      summary: Soft-delete a link
      description: Mark a link inactive and invalidate redirect cache entries.
      operationId: delete_link_api_v1_links__link_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: link_id
        in: path
        required: true
        schema:
          type: integer
          title: Link Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkResponse'
        '400':
          description: Invalid alias, URL, expiration, or request payload.
        '401':
          description: Authentication is required.
        '403':
          description: The current user cannot access this link/workspace.
        '404':
          description: Link not found.
        '409':
          description: Custom alias already exists.
        '410':
          description: The public link is inactive or expired.
        '422':
          description: Request validation failed.
        '429':
          description: Rate limit exceeded.
...
components:
  schemas:
...
    BulkLinkCreateRequest:
      properties:
        urls:
          items:
            type: string
          type: array
          maxItems: 50
          minItems: 1
          title: Urls
        workspace_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Workspace Id
        strip_tracking_params:
          type: boolean
          title: Strip Tracking Params
          default: true
      additionalProperties: false
      type: object
      required:
      - urls
      title: BulkLinkCreateRequest
      description: Bulk URL shortening request.
      example:
        strip_tracking_params: true
        urls:
        - https://example.com/a
        - https://example.com/b?utm_campaign=launch
    BulkLinkCreateResponse:
      properties:
        results:
          items:
            $ref: '#/components/schemas/BulkLinkResult'
          type: array
          title: Results
      type: object
      required:
      - results
      title: BulkLinkCreateResponse
      description: Bulk URL shortening response.
    BulkLinkResult:
      properties:
        index:
          type: integer
          title: Index
        status:
          type: string
          title: Status
        url:
          type: string
          title: Url
        link:
          anyOf:
          - $ref: '#/components/schemas/LinkResponse'
          - type: 'null'
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
      type: object
      required:
      - index
      - status
      - url
      title: BulkLinkResult
      description: Per-item bulk shortening result.
...
    LinkCreate:
      properties:
        destination_url:
          type: string
          maxLength: 4096
          minLength: 1
          title: Destination Url
        workspace_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Workspace Id
        title:
          anyOf:
          - type: string
            maxLength: 255
          - type: 'null'
          title: Title
        custom_alias:
          anyOf:
          - type: string
            maxLength: 32
            minLength: 3
          - type: 'null'
          title: Custom Alias
        password:
          anyOf:
          - type: string
            maxLength: 128
            minLength: 8
          - type: 'null'
          title: Password
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
        strip_tracking_params:
          type: boolean
          title: Strip Tracking Params
          default: true
      additionalProperties: false
      type: object
      required:
      - destination_url
      title: LinkCreate
      description: Request body for creating a short link.
      example:
        custom_alias: example-docs
        destination_url: https://example.com/docs?utm_source=newsletter
        expires_at: '2026-12-31T23:59:59Z'
        password: optional-secret
        strip_tracking_params: true
        title: Example docs
    LinkResponse:
      properties:
        id:
          type: integer
          title: Id
        workspace_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Workspace Id
        owner_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Owner Id
        short_code:
          type: string
          title: Short Code
        destination_url:
          type: string
          title: Destination Url
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        is_password_protected:
          type: boolean
          title: Is Password Protected
        is_active:
          type: boolean
          title: Is Active
        flagged_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Flagged Reason
        checked_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Checked At
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - workspace_id
      - owner_id
      - short_code
      - destination_url
      - title
      - is_password_protected
      - is_active
      - flagged_reason
      - checked_at
      - expires_at
      - created_at
      - updated_at
      title: LinkResponse
      description: Response payload for link CRUD endpoints.
      example:
        checked_at: '2026-06-15T12:00:00Z'
        created_at: '2026-06-15T12:00:00Z'
        destination_url: https://example.com/docs
        id: 742910234935296
        is_active: true
        is_password_protected: false
        owner_id: d7b5f7a2-8a6c-4a46-9443-6480a4f7aa1e
        short_code: 2F9xYz
        title: Example docs
        updated_at: '2026-06-15T12:00:00Z'
    LinkUpdate:
      properties:
        destination_url:
          anyOf:
          - type: string
            maxLength: 4096
            minLength: 1
          - type: 'null'
          title: Destination Url
        title:
          anyOf:
          - type: string
            maxLength: 255
          - type: 'null'
          title: Title
        custom_alias:
          anyOf:
          - type: string
            maxLength: 32
            minLength: 3
          - type: 'null'
          title: Custom Alias
        password:
          anyOf:
          - type: string
            maxLength: 128
            minLength: 8
          - type: 'null'
          title: Password
        clear_password:
          type: boolean
          title: Clear Password
          default: false
        is_active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Active
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
        strip_tracking_params:
          type: boolean
          title: Strip Tracking Params
          default: true
      additionalProperties: false
      type: object
      title: LinkUpdate
      description: Request body for updating link metadata or destination.
      example:
        expires_at: '2027-01-31T23:59:59Z'
        is_active: true
        title: Updated docs link
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login